Moves OnDataLoadedCallback outside the BaseDataManager
